feat: darwin code removed from linux header
